backend/pyproject.toml
[project]
name = "planfirst-backend"
version = "0.1.0"
description = "Planfirst API - Constraint-first travel planning backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pwdlib[argon2]>=0.3.0",  # Modern password hashing with Argon2
    # Pydantic and validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    # OpenAI and AI
    "openai>=1.12.0",
    # Web search
    "ddgs>=8.0.0",
    "requests>=2.31.0",
    "tavily-python>=0.3.0",
    # Environment
    "python-dotenv>=1.0.0",
    # SSE streaming
    "sse-starlette>=1.8.0",
    "aiosqlite>=0.22.1",
    "psycopg[binary]>=3.3.2",
    "greenlet>=3.3.0",
    "pyjwt>=2.10.1",
    "authlib>=1.6.6",
    "itsdangerous>=2.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=1.0.8",
    # Caching
    "diskcache>=5.6.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
